Fixing a warning.

+++ Skin_surface_3/test/Skin_surface_3/union_of_balls_test.cpp
@@ -16,9 +16,9 @@ int main(int, char *[]) {
   typedef CGAL::Exact_predicates_inexact_constructions_kernel IKernel;
   typedef IKernel::Point_3                                     Bare_point;
   typedef CGAL::Weighted_point<Bare_point,IKernel::RT>         Weighted_point;
-  int size=sizeof(pts)/(3*sizeof(double));
+  size_t size=sizeof(pts)/(3*sizeof(double));
   std::vector<Weighted_point> l(size);
-  for (unsigned int i=0; i< size; ++i) {
+  for (size_t i=0; i< size; ++i) {
     l[i]= Weighted_point(Bare_point(pts[i][0], pts[i][1], pts[i][2]),
                          .9*.9);
     std::cout << ".color " << i << std::endl;
